+++ src/functionalTest/java/uk/gov/hmcts/reform/bulkscan/payment/processor/helper/ExceptionRecordCreator.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.bulkscan.orchestrator.helper;
+package uk.gov.hmcts.reform.bulkscan.payment.processor.helper;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
